requirements.txt
# Flutter App Factory Requirements

# Core dependencies
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

# CLI and interface
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Marketing automation
google-play-scraper>=1.2.0
app-store-scraper>=0.3.5
openai>=1.0.0
beautifulsoup4>=4.12.0

# Deployment
fastlane>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Monitoring
sentry-sdk>=1.0.0
prometheus-client>=0.17.0

# Async support
asyncio>=3.4.3
aiohttp>=3.8.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# File system monitoring
watchdog>=3.0.0

# MCP Integration
supabase>=2.0.0
GitPython>=3.1.0

# Database (for tracking)
sqlalchemy>=2.0.0
alembic>=1.11.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3